requirements.txt
python-telegram-bot==20.7
python-dotenv
openai
pymongo
apscheduler
requests